Updated manifest.json
Updated manifest file to adhere more to google's recommended set up for the manifest file. Short name was also added to manifest file as well as the icons object. Once the icons are added with the appropriate sizes, the changes will be pushed with the new version 0.1.1 which was added in this update.